refactor: major code quality improvements for v3.0.1
This release represents a significant internal refactoring while maintaining 100% backward compatibility.

**Task Store Modularization (78% reduction)**
- Refactored taskStore.ts from 879 to 190 lines
- Extracted logic into 6 focused modules following Single Responsibility Principle:
  - taskStoreHelpers.ts - Helper functions and filter utilities
  - taskStoreActions.ts - CRUD operations (add, update, delete, move, archive)
  - taskStoreFilters.ts - Search and filtering with caching
  - taskStoreSearch.ts - Search navigation and preferences
  - taskStoreImportExport.ts - Import/export operations
  - taskStoreValidation.ts - Validation and error handling

**Export/Import Simplification**
- Reduced detectImportConflicts from 64 to 20 lines (70% reduction)
- Reduced processImportData from 68 to 34 lines (50% reduction)
- Created exportImportHelpers.ts with 9 focused helper functions

**Dead Code Removal**
- Removed deploymentValidator.ts (724 lines of unused code)
- Deleted unused modular stores (taskCrudStore, taskDataStore, taskSearchStore)
- Applied YAGNI principle throughout

**Code Quality Improvements**
- All functions now under 30 lines for better readability
- Consistent application of DRY principle
- Enhanced testability through modular architecture

**Documentation Updates**
- Created CHANGELOG.md documenting v3.0.1 changes
- Created docs/REFACTORING-V3.md comprehensive refactoring guide
- Updated README.md with v3.0.1 architecture details
- Updated docs/developer-guide.md with modular store patterns
- Updated package.json to version 3.0.1
- Updated sample-export.json with current dates

**Impact Summary**
- Total: 12 files changed, 457 insertions(+), 2265 deletions(-)
- taskStore.ts: 879 → 190 lines (78% reduction)
- Longest function: 70 → <30 lines (57%+ reduction)
- Dead code: 724 → 0 lines (100% removal)
- Store modules: 1 → 7 focused files
