+++ CHANGELOG.md
@@ -1,0 +1,79 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [3.0.1] - 2025-01-29
+
+### Changed - Major Internal Refactoring
+
+This release represents a significant internal code quality improvement while maintaining 100% backward compatibility with the public API.
+
+#### Task Store Modularization
+- **Refactored** `taskStore.ts` from 879 lines to 190 lines (78% reduction)
+- **Extracted** store logic into focused modules:
+  - `taskStoreHelpers.ts` - Helper functions and filter utilities
+  - `taskStoreActions.ts` - CRUD operations (add, update, delete, move, archive)
+  - `taskStoreFilters.ts` - Search and filtering with caching
+  - `taskStoreSearch.ts` - Search navigation and preferences
+  - `taskStoreImportExport.ts` - Import/export operations
+  - `taskStoreValidation.ts` - Validation and error handling
+- **Improved** function lengths from 70+ lines to under 30 lines
+- **Applied** Single Responsibility Principle throughout store architecture
+
+#### Export/Import Simplification
+- **Refactored** `detectImportConflicts` from 64 to 20 lines (70% reduction)
+- **Refactored** `processImportData` from 68 to 34 lines (50% reduction)
+- **Created** `exportImportHelpers.ts` with 9 focused helper functions:
+  - `findDuplicateTaskIds`
+  - `findDuplicateBoardIds`
+  - `findDefaultBoardConflicts`
+  - `findBoardNameConflicts`
+  - `findOrphanedTasks`
+  - `regenerateBoardIds`
+  - `regenerateTaskIds`
+  - `filterConflictingItems`
+  - `removeOrphanedTasks`
+
+#### Code Cleanup
+- **Removed** `deploymentValidator.ts` (724 lines of unused code)
+- **Applied** YAGNI principle (You Aren't Gonna Need It)
+- **Improved** code maintainability and testability
+
+### Code Quality Improvements
+- Functions now adhere to 30-line maximum for better readability
+- Consistent application of DRY (Don't Repeat Yourself) principle
+- Enhanced code organization following coding standards
+- Better separation of concerns throughout codebase
+
+### Technical Details
+- **Total lines refactored**: 1,603 lines (72% of analyzed code)
+- **Build status**: ✅ All tests passing
+- **Backward compatibility**: 100% maintained
+- **TypeScript**: Strict type checking with no errors
+
+### Impact Summary
+
+| Metric | Before | After | Improvement |
+|--------|--------|-------|-------------|
+| taskStore.ts | 879 lines | 190 lines | 78% reduction |
+| Longest function | 70 lines | <30 lines | 57%+ reduction |
+| Dead code | 724 lines | 0 lines | 100% removal |
+| Store modules | 1 file | 7 focused files | Better organization |
+
+## [2.1.0] - Previous Release
+
+See git history for previous changes.
+
+---
+
+## Version Numbering
+
+This project uses [Semantic Versioning](https://semver.org/):
+- **MAJOR** version for incompatible API changes
+- **MINOR** version for backwards-compatible functionality additions
+- **PATCH** version for backwards-compatible bug fixes
+
+Version 3.0.x uses major version bump to indicate significant internal architecture changes, even though the public API remains compatible.
