+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kanban-todos-temp",
-  "version": "2.1.0",
+  "version": "3.0.1",
   "private": true,
   "scripts": {
     "dev": "next dev",
