--- src/lib/utils/deploymentValidator.ts
@@ -1,724 +1,0 @@
-/**
- * Production Deployment Validator
- * Validates deployment configuration and environment setup
- */
-
-export interface DeploymentCheck {
-  id: string;
-  name: string;
-  description: string;
-  category: 'environment' | 'configuration' | 'dependencies' | 'build' | 'deployment';
-  status: 'pass' | 'fail' | 'warning' | 'info';
-  message: string;
-  recommendation?: string;
-  autoFixable: boolean;
-  fixAction?: () => Promise<void>;
-}
-
-export interface DeploymentReport {
-  overallStatus: 'ready' | 'not-ready' | 'warning';
-  totalChecks: number;
-  passedChecks: number;
-  failedChecks: number;
-  warningChecks: number;
-  categories: Record<string, DeploymentCheck[]>;
-  criticalIssues: string[];
-  recommendations: string[];
-  environmentInfo: {
-    nodeVersion: string;
-    npmVersion: string;
-    buildMode: string;
-    environment: string;
-  };
-}
-
-export class DeploymentValidator {
-  private static instance: DeploymentValidator;
-  private checks: DeploymentCheck[] = [];
-
-  constructor() {
-    this.initializeChecks();
-  }
-
-  static getInstance(): DeploymentValidator {
-    if (!this.instance) {
-      this.instance = new DeploymentValidator();
-    }
-    return this.instance;
-  }
-
-  private initializeChecks(): void {
-    this.checks = [
-      // Environment Checks
-      {
-        id: 'node-version',
-        name: 'Node.js Version',
-        description: 'Verify Node.js version compatibility',
-        category: 'environment',
-        status: 'info',
-        message: 'Checking Node.js version...',
-        autoFixable: false,
-      },
-      {
-        id: 'npm-version',
-        name: 'NPM Version',
-        description: 'Verify NPM version compatibility',
-        category: 'environment',
-        status: 'info',
-        message: 'Checking NPM version...',
-        autoFixable: false,
-      },
-      {
-        id: 'environment-variables',
-        name: 'Environment Variables',
-        description: 'Verify required environment variables are set',
-        category: 'environment',
-        status: 'info',
-        message: 'Checking environment variables...',
-        autoFixable: false,
-      },
-      {
-        id: 'build-mode',
-        name: 'Build Mode',
-        description: 'Verify build is in production mode',
-        category: 'environment',
-        status: 'info',
-        message: 'Checking build mode...',
-        autoFixable: false,
-      },
-
-      // Configuration Checks
-      {
-        id: 'next-config',
-        name: 'Next.js Configuration',
-        description: 'Verify Next.js configuration is production-ready',
-        category: 'configuration',
-        status: 'info',
-        message: 'Checking Next.js configuration...',
-        autoFixable: false,
-      },
-      {
-        id: 'typescript-config',
-        name: 'TypeScript Configuration',
-        description: 'Verify TypeScript configuration is optimized',
-        category: 'configuration',
-        status: 'info',
-        message: 'Checking TypeScript configuration...',
-        autoFixable: false,
-      },
-      {
-        id: 'tailwind-config',
-        name: 'Tailwind Configuration',
-        description: 'Verify Tailwind CSS configuration is optimized',
-        category: 'configuration',
-        status: 'info',
-        message: 'Checking Tailwind configuration...',
-        autoFixable: false,
-      },
-      {
-        id: 'eslint-config',
-        name: 'ESLint Configuration',
-        description: 'Verify ESLint configuration is production-ready',
-        category: 'configuration',
-        status: 'info',
-        message: 'Checking ESLint configuration...',
-        autoFixable: false,
-      },
-
-      // Dependencies Checks
-      {
-        id: 'dependencies',
-        name: 'Dependencies',
-        description: 'Verify all dependencies are properly installed',
-        category: 'dependencies',
-        status: 'info',
-        message: 'Checking dependencies...',
-        autoFixable: false,
-      },
-      {
-        id: 'security-audit',
-        name: 'Security Audit',
-        description: 'Verify no security vulnerabilities in dependencies',
-        category: 'dependencies',
-        status: 'info',
-        message: 'Checking security vulnerabilities...',
-        autoFixable: false,
-      },
-      {
-        id: 'outdated-packages',
-        name: 'Outdated Packages',
-        description: 'Check for outdated packages',
-        category: 'dependencies',
-        status: 'info',
-        message: 'Checking for outdated packages...',
-        autoFixable: false,
-      },
-
-      // Build Checks
-      {
-        id: 'build-success',
-        name: 'Build Success',
-        description: 'Verify build completes successfully',
-        category: 'build',
-        status: 'info',
-        message: 'Checking build success...',
-        autoFixable: false,
-      },
-      {
-        id: 'bundle-analysis',
-        name: 'Bundle Analysis',
-        description: 'Analyze bundle size and composition',
-        category: 'build',
-        status: 'info',
-        message: 'Analyzing bundle...',
-        autoFixable: false,
-      },
-      {
-        id: 'static-export',
-        name: 'Static Export',
-        description: 'Verify static export is properly configured',
-        category: 'build',
-        status: 'info',
-        message: 'Checking static export...',
-        autoFixable: false,
-      },
-      {
-        id: 'image-optimization',
-        name: 'Image Optimization',
-        description: 'Verify images are optimized for production',
-        category: 'build',
-        status: 'info',
-        message: 'Checking image optimization...',
-        autoFixable: false,
-      },
-
-      // Deployment Checks
-      {
-        id: 'deployment-config',
-        name: 'Deployment Configuration',
-        description: 'Verify deployment configuration is complete',
-        category: 'deployment',
-        status: 'info',
-        message: 'Checking deployment configuration...',
-        autoFixable: false,
-      },
-      {
-        id: 'cdn-config',
-        name: 'CDN Configuration',
-        description: 'Verify CDN configuration is optimized',
-        category: 'deployment',
-        status: 'info',
-        message: 'Checking CDN configuration...',
-        autoFixable: false,
-      },
-      {
-        id: 'ssl-certificate',
-        name: 'SSL Certificate',
-        description: 'Verify SSL certificate is valid',
-        category: 'deployment',
-        status: 'info',
-        message: 'Checking SSL certificate...',
-        autoFixable: false,
-      },
-      {
-        id: 'domain-config',
-        name: 'Domain Configuration',
-        description: 'Verify domain configuration is correct',
-        category: 'deployment',
-        status: 'info',
-        message: 'Checking domain configuration...',
-        autoFixable: false,
-      },
-    ];
-  }
-
-  async validateDeployment(): Promise<DeploymentReport> {
-    const results: DeploymentCheck[] = [];
-    
-    for (const check of this.checks) {
-      try {
-        const result = await this.runCheck(check);
-        results.push(result);
-      } catch {
-        results.push({
-          ...check,
-          status: 'fail',
-          message: 'Check failed: Unknown error',
-        });
-      }
-    }
-
-    return this.generateReport(results);
-  }
-
-  private async runCheck(check: DeploymentCheck): Promise<DeploymentCheck> {
-    switch (check.id) {
-      case 'node-version':
-        return await this.checkNodeVersion(check);
-      case 'npm-version':
-        return await this.checkNPMVersion(check);
-      case 'environment-variables':
-        return await this.checkEnvironmentVariables(check);
-      case 'build-mode':
-        return await this.checkBuildMode(check);
-      case 'next-config':
-        return await this.checkNextConfig(check);
-      case 'typescript-config':
-        return await this.checkTypeScriptConfig(check);
-      case 'tailwind-config':
-        return await this.checkTailwindConfig(check);
-      case 'eslint-config':
-        return await this.checkESLintConfig(check);
-      case 'dependencies':
-        return await this.checkDependencies(check);
-      case 'security-audit':
-        return await this.checkSecurityAudit(check);
-      case 'outdated-packages':
-        return await this.checkOutdatedPackages(check);
-      case 'build-success':
-        return await this.checkBuildSuccess(check);
-      case 'bundle-analysis':
-        return await this.checkBundleAnalysis(check);
-      case 'static-export':
-        return await this.checkStaticExport(check);
-      case 'image-optimization':
-        return await this.checkImageOptimization(check);
-      case 'deployment-config':
-        return await this.checkDeploymentConfig(check);
-      case 'cdn-config':
-        return await this.checkCDNConfig(check);
-      case 'ssl-certificate':
-        return await this.checkSSLCertificate(check);
-      case 'domain-config':
-        return await this.checkDomainConfig(check);
-      default:
-        return {
-          ...check,
-          status: 'fail',
-          message: 'Unknown check type',
-        };
-    }
-  }
-
-  private async checkNodeVersion(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      const nodeVersion = process.version;
-      const majorVersion = parseInt(nodeVersion.split('.')[0].substring(1));
-      
-      if (majorVersion >= 18) {
-        return {
-          ...check,
-          status: 'pass',
-          message: `Node.js version ${nodeVersion} is compatible`,
-        };
-      } else {
-        return {
-          ...check,
-          status: 'fail',
-          message: `Node.js version ${nodeVersion} is not compatible. Requires Node.js 18+`,
-          recommendation: 'Upgrade to Node.js 18 or higher',
-        };
-      }
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not determine Node.js version',
-      };
-    }
-  }
-
-  private async checkNPMVersion(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      const npmVersion = process.env.npm_version || 'unknown';
-      
-      return {
-        ...check,
-        status: 'pass',
-        message: `NPM version ${npmVersion} is compatible`,
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not determine NPM version',
-      };
-    }
-  }
-
-  private async checkEnvironmentVariables(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      const requiredVars = ['NODE_ENV'];
-      const missingVars = requiredVars.filter(varName => !process.env[varName]);
-      
-      if (missingVars.length === 0) {
-        return {
-          ...check,
-          status: 'pass',
-          message: 'All required environment variables are set',
-        };
-      } else {
-        return {
-          ...check,
-          status: 'fail',
-          message: `Missing environment variables: ${missingVars.join(', ')}`,
-          recommendation: 'Set all required environment variables',
-        };
-      }
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not check environment variables',
-      };
-    }
-  }
-
-  private async checkBuildMode(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      const isProduction = process.env.NODE_ENV === 'production';
-      
-      if (isProduction) {
-        return {
-          ...check,
-          status: 'pass',
-          message: 'Build is in production mode',
-        };
-      } else {
-        return {
-          ...check,
-          status: 'warning',
-          message: 'Build is not in production mode',
-          recommendation: 'Set NODE_ENV to production for deployment',
-        };
-      }
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not check build mode',
-      };
-    }
-  }
-
-  private async checkNextConfig(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check if next.config.ts exists and is properly configured
-      return {
-        ...check,
-        status: 'pass',
-        message: 'Next.js configuration is properly set up',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not verify Next.js configuration',
-      };
-    }
-  }
-
-  private async checkTypeScriptConfig(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check if tsconfig.json exists and is properly configured
-      return {
-        ...check,
-        status: 'pass',
-        message: 'TypeScript configuration is properly set up',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not verify TypeScript configuration',
-      };
-    }
-  }
-
-  private async checkTailwindConfig(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check if tailwind.config.js exists and is properly configured
-      return {
-        ...check,
-        status: 'pass',
-        message: 'Tailwind CSS configuration is properly set up',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not verify Tailwind configuration',
-      };
-    }
-  }
-
-  private async checkESLintConfig(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check if eslint.config.mjs exists and is properly configured
-      return {
-        ...check,
-        status: 'pass',
-        message: 'ESLint configuration is properly set up',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not verify ESLint configuration',
-      };
-    }
-  }
-
-  private async checkDependencies(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check if all dependencies are properly installed
-      return {
-        ...check,
-        status: 'pass',
-        message: 'All dependencies are properly installed',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not verify dependencies',
-      };
-    }
-  }
-
-  private async checkSecurityAudit(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would run npm audit to check for security vulnerabilities
-      return {
-        ...check,
-        status: 'pass',
-        message: 'No security vulnerabilities found in dependencies',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not run security audit',
-      };
-    }
-  }
-
-  private async checkOutdatedPackages(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check for outdated packages
-      return {
-        ...check,
-        status: 'info',
-        message: 'Package versions are up to date',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not check for outdated packages',
-      };
-    }
-  }
-
-  private async checkBuildSuccess(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check if the build was successful
-      return {
-        ...check,
-        status: 'pass',
-        message: 'Build completed successfully',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Build failed or could not be verified',
-      };
-    }
-  }
-
-  private async checkBundleAnalysis(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would analyze the bundle size and composition
-      return {
-        ...check,
-        status: 'pass',
-        message: 'Bundle analysis completed successfully',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not analyze bundle',
-      };
-    }
-  }
-
-  private async checkStaticExport(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check if static export is properly configured
-      return {
-        ...check,
-        status: 'pass',
-        message: 'Static export is properly configured',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'fail',
-        message: 'Could not verify static export configuration',
-      };
-    }
-  }
-
-  private async checkImageOptimization(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check if images are optimized
-      return {
-        ...check,
-        status: 'pass',
-        message: 'Images are optimized for production',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not verify image optimization',
-      };
-    }
-  }
-
-  private async checkDeploymentConfig(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check deployment configuration files
-      return {
-        ...check,
-        status: 'pass',
-        message: 'Deployment configuration is complete',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not verify deployment configuration',
-      };
-    }
-  }
-
-  private async checkCDNConfig(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check CDN configuration
-      return {
-        ...check,
-        status: 'pass',
-        message: 'CDN configuration is optimized',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not verify CDN configuration',
-      };
-    }
-  }
-
-  private async checkSSLCertificate(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check SSL certificate validity
-      return {
-        ...check,
-        status: 'pass',
-        message: 'SSL certificate is valid',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not verify SSL certificate',
-      };
-    }
-  }
-
-  private async checkDomainConfig(check: DeploymentCheck): Promise<DeploymentCheck> {
-    try {
-      // This would check domain configuration
-      return {
-        ...check,
-        status: 'pass',
-        message: 'Domain configuration is correct',
-      };
-    } catch {
-      return {
-        ...check,
-        status: 'warning',
-        message: 'Could not verify domain configuration',
-      };
-    }
-  }
-
-  private generateReport(results: DeploymentCheck[]): DeploymentReport {
-    const categories: Record<string, DeploymentCheck[]> = {};
-    
-    // Group checks by category
-    results.forEach(check => {
-      if (!categories[check.category]) {
-        categories[check.category] = [];
-      }
-      categories[check.category].push(check);
-    });
-
-    const totalChecks = results.length;
-    const passedChecks = results.filter(c => c.status === 'pass').length;
-    const failedChecks = results.filter(c => c.status === 'fail').length;
-    const warningChecks = results.filter(c => c.status === 'warning').length;
-
-    // Determine overall status
-    let overallStatus: 'ready' | 'not-ready' | 'warning' = 'ready';
-    if (failedChecks > 0) {
-      overallStatus = 'not-ready';
-    } else if (warningChecks > 0) {
-      overallStatus = 'warning';
-    }
-
-    // Collect critical issues and recommendations
-    const criticalIssues: string[] = [];
-    const recommendations: string[] = [];
-    
-    results.forEach(check => {
-      if (check.status === 'fail') {
-        criticalIssues.push(`${check.name}: ${check.message}`);
-        if (check.recommendation) {
-          recommendations.push(check.recommendation);
-        }
-      }
-    });
-
-    // Environment info
-    const environmentInfo = {
-      nodeVersion: process.version,
-      npmVersion: process.env.npm_version || 'unknown',
-      buildMode: process.env.NODE_ENV || 'development',
-      environment: process.env.NODE_ENV || 'development',
-    };
-
-    return {
-      overallStatus,
-      totalChecks,
-      passedChecks,
-      failedChecks,
-      warningChecks,
-      categories,
-      criticalIssues,
-      recommendations,
-      environmentInfo,
-    };
-  }
-
-  getChecks(): DeploymentCheck[] {
-    return [...this.checks];
-  }
-
-  getCheckById(id: string): DeploymentCheck | undefined {
-    return this.checks.find(check => check.id === id);
-  }
-}
